rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG application for SQL generation"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.0.0",
    "numpy>=1.20.0",
    "python-multipart",
    "sqlglot>=28.0.0",
    "python-multipart",
    "psycopg>=3.3.2",
    "psycopg-binary==3.3.2",
    "psycopg-pool==3.3.0",
    "rank-bm25>=0.2.2",
    "fastmcp>=2.14.0",
    "fastapi>=0.116.1",
    "PyMuPDF>=1.23.0", # для PDF
    "python-docx>=1.1.0", # для DOCX
    "yandexcloud>=0.371.0",
]

[project.scripts]
rag = "main:main"
rag-run = "main:run_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
